+++ CHANGELOG.md
@@ -4,9 +4,15 @@ All notable changes to this project will be documented in this file.
 
 ## 1.17.0
 
+### Upgrade Notes
+
+- CRAIG no longer depends on react-scripts and uses Next JS to build
+- Updated VPC Cluster and Worker Pool resource addressess to be more intuitive
+
 ### Features
 
 - When using dynamic subnets, CIDR blocks are now calculated to use the fewest total IPs within a VPC
+- Security and Compliance Center (SCC) has been removed as the service is no longer supported on IBM Cloud
 
 ## 1.16.5